Restructure unit test naming scheme.
